pyproject.toml
[project]
name = "chellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-ollama>=0.3.5",
    "python-chess>=1.999",
]
